deps/v8/test/inspector/runtime/stable-object-id-expected.txt

Checks that protocol returns the same RemoteObjectId for the same object

Running test: testGlobal
Compare global evaluated twice: true

Running test: testObject
Compare object evaluated twice: true

Running test: testObjectInArray
Compare first and second element: true

Running test: testObjectOnPause
Compare global and this: true
Compare global and global on pause: true
Compare a and a on pause: true
